drm: Reduce latency while processing atomic ioctls
Unfortunately, our display driver must occasionally sleep inside
start_atomic(), which causes the CPU running the ioctl to schedule and
potentially idle. Depending on how deeply the running CPU idles, there
can be quite a bit of latency added to processing the "atomic" ioctl as
a result, which hurts display rendering latency. We can alleviate this
by optimistically assuming that the task running the ioctl won't
migrate CPUs while it runs and restricting the current CPU to shallow
idle states.

Change-Id: Ia4e19ff31dde75bb716f6b35d37c4f5dc1cf58e7
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>